backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.18
aiofiles==24.1.0
pytest==8.3.4
